move SoundBuffer to accessor folder
